+++ test/ecto/query/builder/order_by_test.exs
@@ -26,6 +26,12 @@ defmodule Ecto.Query.Builder.OrderByTest do
       assert {Macro.escape(quote do [asc: &0.x, desc: &0.y] end), %{}} ==
              escape(:order_by, quote do [:x, desc: :y] end, [x: 0], __ENV__)
 
+      assert {Macro.escape(quote do [asc_nulls_first: &0.x, desc_nulls_first: &0.y] end), %{}} ==
+             escape(:order_by, quote do [asc_nulls_first: :x, desc_nulls_first: :y] end, [x: 0], __ENV__)
+
+      assert {Macro.escape(quote do [asc_nulls_last: &0.x, desc_nulls_last: &0.y] end), %{}} ==
+             escape(:order_by, quote do [asc_nulls_last: :x, desc_nulls_last: :y] end, [x: 0], __ENV__)
+
       import Kernel, except: [>: 2]
       assert {Macro.escape(quote do [asc: 1 > 2] end), %{}} ==
              escape(:order_by, quote do 1 > 2 end, [], __ENV__)
@@ -38,7 +44,7 @@ defmodule Ecto.Query.Builder.OrderByTest do
     end
 
     test "raises on unknown expression" do
-      message = "expected :asc, :desc or interpolated value in `order_by`, got: `:test`"
+      message = ~r":desc_nulls_first or interpolated value in `order_by`, got: `:test`"
       assert_raise Ecto.Query.CompileError, message, fn ->
         escape(:order_by, quote do [test: x.y] end, [x: 0], __ENV__)
       end
@@ -56,8 +62,7 @@ defmodule Ecto.Query.Builder.OrderByTest do
     end
 
     test "raises on invalid direction" do
-      message = "expected :asc or :desc in `order_by`, got: `:temp`"
-      assert_raise Ecto.Query.CompileError, message, fn ->
+      assert_raise Ecto.Query.CompileError, ~r"expected one of :asc,", fn ->
         temp = :temp
         order_by("posts", [p], [{^var!(temp), p.y}])
       end